pyproject.toml
[project]
name = "uas-researcher"
version = "0.1.0"
description = "Multi-agent UAS research system using Claude with web search"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
keywords = ["uas", "drone", "multi-agent", "claude", "research", "anthropic"]
authors = [{ name = "Arthur Fantaci" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "pre-commit>=4.0.0",
]

[project.urls]
Repository = "https://github.com/arthurfantaci/uas-researcher"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents"]

# =============================================================================
# Ruff Configuration
# =============================================================================
[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "D",      # pydocstyle
    "UP",     # pyupgrade
    "ANN",    # flake8-annotations
    "S",      # flake8-bandit (security)
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "RUF",    # Ruff-specific rules
    "TRY",    # tryceratops (exception handling)
    "EM",     # flake8-errmsg (exception messages)
    "PIE",    # flake8-pie (misc linting)
    "RET",    # flake8-return
    "ARG",    # flake8-unused-arguments
    "PL",     # Pylint
]

ignore = [
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package
    "D107",   # Missing docstring in __init__
    "ANN401", # Dynamically typed expressions (Any)
    "S101",   # Use of assert (needed for tests)
    "TRY003", # Long exception messages
    "TRY300", # try-consider-else
    "TRY400", # error-instead-of-exception
    "TRY401", # verbose-log-message
    "EM101",  # String literal in exception
    "EM102",  # f-string in exception
    "RET504", # Unnecessary variable before return
    "PLR0913",# Too many function arguments
    "PLR0911",# Too many return statements
    "PLR0912",# Too many branches
    "PLR0915",# Too many statements
    "PLR2004",# Magic value in comparison
    "PLC0415",# import-outside-top-level
    "PLW0603",# global-statement
    "ARG001", # unused-function-argument
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # Allow assert in tests
    "ANN",     # Don't require annotations in tests
    "D",       # Don't require docstrings in tests
    "PLR2004", # Magic values in test assertions
    "ARG",     # Unused fixture args common in pytest
]
"agents/specialists.py" = ["E501", "RUF001"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# =============================================================================
# Pytest Configuration
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = ["-ra", "-q"]
